Reduce redux saveState overhead
I noticed my CPU was quite high when downloading big streams and most of
the CPU was being spent doing browser.storage.local.set(...) for saving
the state.  This state only appears to be needed when the background
page is reloaded (which will never happen if the a content-script is
active.

Buffering the save by 200ms reduces JS overhead by about ~40% (not
including native methods overhead) and total CPU usage on my system
(circa 2022 AMD 16 core) is reduced by ~28%.